% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster-package.R
\name{assign_possession_ids}
\alias{assign_possession_ids}
\title{assign id to each possession, which can be used to calculate posessions
for each lineup.}
\usage{
assign_possession_ids(pbp_df)
}
\arguments{
\item{pbp_df}{the play by play data.frame}
}
\value{
data.frame that mirrors pbp_df but with an additional possession_id column
}
\description{
assign id to each possession, which can be used to calculate posessions
for each lineup.
}
