% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster-package.R
\docType{data}
\name{pbp}
\alias{pbp}
\title{Play-by-play data}
\format{A data frame with 525 rows and 75 variables}
\description{
Play by play data from a game between the Memphis Grizzlies and Sacramento
Kings on 1 January 2017.
}
