% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster-package.R
\docType{data}
\name{players}
\alias{players}
\title{Player data}
\format{a data frame with 23 rows and 6 variables}
\description{
Player information for game participants
}
\details{
\itemize{
  \item player_id
  \item full_name
  \item team_id
  \item team_market
  \item team_name
  \item season
}
}
