% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster-package.R
\name{plus_minus}
\alias{plus_minus}
\title{Calculate PM for each lineup in a single game}
\usage{
plus_minus(pbp_df)
}
\arguments{
\item{pbp_df}{the play by play data.frame}
}
\value{
data.frame
}
\description{
Calculate PM for each lineup in a single game
}
