% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster-package.R
\name{sort_pbp}
\alias{sort_pbp}
\title{make sure play by play is sorted first by period, then by clock}
\usage{
sort_pbp(pbp_df)
}
\value{
data.frame
}
\description{
make sure play by play is sorted first by period, then by clock
}
